feat: add consolidated HTTP types
