+++ backend/controllers/AssetController.ts
@@ -1,5 +1,4 @@
-import { AuthedRequest } from '../types/AuthedRequest';
-import { AuthedRequestHandler } from '../types/AuthedRequestHandler';
+import { AuthedRequest, AuthedRequestHandler } from '../types/http';
 import Asset from '../models/Asset';
 import mongoose from 'mongoose';
 import { validationResult } from 'express-validator';
